# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NutriBasket API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile and goals
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: No profile set up yet
    put:
      operationId: upsertProfile
      tags: [profile]
      summary: Create or update user profile and goals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /recipes:
    get:
      operationId: listRecipes
      tags: [recipes]
      summary: List/search recipes with filters
      parameters:
        - name: query
          in: query
          schema:
            type: string
        - name: goal
          in: query
          schema:
            type: string
            enum: [high_protein, low_calorie, muscle_gain, fat_loss, budget, vegan, low_carb]
        - name: maxPrepTime
          in: query
          schema:
            type: integer
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [easy, medium, hard]
        - name: maxCaloriesPerServing
          in: query
          schema:
            type: integer
        - name: minProteinPerServing
          in: query
          schema:
            type: integer
        - name: maxCost
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
  /recipes/{id}:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get a single recipe with full ingredient details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recipe detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeDetail"
        "404":
          description: Not found
...
  /saved/recipes:
    get:
      operationId: listSavedRecipes
      tags: [saved]
      summary: List saved/favourite recipes
      responses:
        "200":
          description: Saved recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
    post:
      operationId: saveRecipe
      tags: [saved]
      summary: Save a recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveItemInput"
      responses:
        "201":
          description: Saved
  /saved/recipes/{recipeId}:
    delete:
      operationId: unsaveRecipe
      tags: [saved]
      summary: Remove a recipe from saved
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /saved/snacks:
    get:
      operationId: listSavedSnacks
      tags: [saved]
      summary: List saved snacks
      responses:
        "200":
          description: Saved snacks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SnackSuggestion"
    post:
      operationId: saveSnack
      tags: [saved]
      summary: Save a snack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveItemInput"
      responses:
        "201":
          description: Saved
  /saved/snacks/{productId}:
    delete:
      operationId: unsaveSnack
      tags: [saved]
      summary: Remove a snack from saved
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, currentWeightKg, heightCm, targetWeightKg, ageYears, sex, activityLevel, dietPreference, budgetWeekly, mealFrequency]
      properties:
        id:
          type: integer
        currentWeightKg:
          type: number
        heightCm:
          type: number
        targetWeightKg:
          type: number
        ageYears:
          type: integer
        sex:
          type: string
          enum: [male, female, other]
        activityLevel:
          type: string
          enum: [sedentary, lightly_active, moderately_active, very_active, extra_active]
        bodyFatPercent:
          type: ["number", "null"]
        dietPreference:
          type: string
          enum: [standard, high_protein, low_calorie, low_carb, vegan, halal, vegetarian]
        budgetWeekly:
          type: number
        mealFrequency:
          type: integer
        retailerPreferences:
          type: array
          items:
            type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserProfileInput:
      type: object
      required: [currentWeightKg, heightCm, targetWeightKg, ageYears, sex, activityLevel, dietPreference, budgetWeekly, mealFrequency]
      properties:
        currentWeightKg:
          type: number
        heightCm:
          type: number
        targetWeightKg:
          type: number
        ageYears:
          type: integer
        sex:
          type: string
          enum: [male, female, other]
        activityLevel:
          type: string
          enum: [sedentary, lightly_active, moderately_active, very_active, extra_active]
        bodyFatPercent:
          type: ["number", "null"]
        dietPreference:
          type: string
          enum: [standard, high_protein, low_calorie, low_carb, vegan, halal, vegetarian]
        budgetWeekly:
          type: number
        mealFrequency:
          type: integer
        retailerPreferences:
          type: array
          items:
            type: integer
...
    Recipe:
      type: object
      required: [id, name, prepTimeMin, servings, caloriesPerServing, proteinPerServingG, carbsPerServingG, fatPerServingG, difficulty, tags, estimatedCost, imageUrl]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        prepTimeMin:
          type: integer
        cookTimeMin:
          type: integer
        servings:
          type: integer
        caloriesPerServing:
          type: integer
        proteinPerServingG:
          type: number
        carbsPerServingG:
          type: number
        fatPerServingG:
          type: number
        fiberPerServingG:
          type: ["number", "null"]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        tags:
          type: array
          items:
            type: string
        estimatedCost:
          type: number
        imageUrl:
          type: string
        isSaved:
          type: boolean
    RecipeDetail:
      type: object
      required: [id, name, prepTimeMin, servings, caloriesPerServing, proteinPerServingG, carbsPerServingG, fatPerServingG, difficulty, tags, estimatedCost, imageUrl, ingredients, instructions]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        prepTimeMin:
          type: integer
        cookTimeMin:
          type: integer
        servings:
          type: integer
        caloriesPerServing:
          type: integer
        proteinPerServingG:
          type: number
        carbsPerServingG:
          type: number
        fatPerServingG:
          type: number
        fiberPerServingG:
          type: ["number", "null"]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        tags:
          type: array
          items:
            type: string
        estimatedCost:
          type: number
        imageUrl:
          type: string
        isSaved:
          type: boolean
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
        instructions:
          type: array
          items:
            type: string
    RecipeIngredient:
      type: object
      required: [id, name, quantity, unit, calories, proteinG, carbsG, fatG, estimatedCost]
      properties:
        id:
          type: integer
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        calories:
          type: integer
        proteinG:
          type: number
        carbsG:
          type: number
        fatG:
          type: number
        estimatedCost:
          type: number
        productId:
          type: ["integer", "null"]
        substitutes:
          type: array
          items:
            type: string

    # ---- Products ----
...
    SnackSuggestion:
      type: object
      required: [productId, name, retailerName, priceAud, caloriesPerServing, proteinPerServingG, sugarPerServingG, fatPerServingG, servingSize, servingUnit, isOnSpecial]
      properties:
        productId:
          type: integer
        name:
          type: string
        retailerName:
          type: string
        priceAud:
          type: number
        caloriesPerServing:
          type: integer
        proteinPerServingG:
          type: number
        sugarPerServingG:
          type: number
        fatPerServingG:
          type: number
        servingSize:
          type: number
        servingUnit:
          type: string
        isOnSpecial:
          type: boolean
        savingsPercent:
          type: ["number", "null"]
        imageUrl:
          type: string
        isSaved:
          type: boolean
...
    SaveItemInput:
      type: object
      required: [itemId]
      properties:
        itemId:
          type: integer